Farmer Drops Keycard; Fixed Purple Textures on Game Objects
Allowed the farmer to drop their keycard when they're stunned.
Player is now fully hidden from the farmer when they're concealed.
Fixed null reference error for a spotlight's pickup point, in the fuse room.
Fixed interactable game-objects having purple textures when the player's out of range to interact with them.